Added mock homescreen
